source/Serpentarium/Source/TaskManager/requirements.txt
django-cors-headers==3.2.1
django==3.0.4
djangorestframework==3.11.0
opencensus-ext-azure==1.0.2
opencensus-ext-django==0.7.2
opencensus-ext-logging==0.1.0
opencensus-ext-requests==0.7.3
wrapt==1.11.*
pylint==2.4.4
pytest-cov==2.8.1
pytest-django==3.8.0
pytest==5.4.1

jsonmerge==1.7.0